- Make the buffer functions take only the grid size and block size - Centralize the automatic scheduling in compyte_kernel - Add properties to support automatic scheduling - Reflect changes in gpuarray.pyx
This is partly because I discovered that OpenCL can be really stupid in its allocations since it will never schedule more threads than the total number.  We can go a bit over for the sake of performance.
Also manual scheduling is required to have a non-crappy reduction performance (which is coming really soon).