Use NVRTC and ptxas to generate cubin.
libgpuarray has been using nvcc to compile CUDA source code to cubin.
This change replace the nvcc invocation with the NVRTC and ptxas invocation.
As nvcc is an offline tool that spawns many child processes, it is much slower
than the new combination.  NVRTC is a library, so there is no overhead
associated with child process management, and ptxas does not spawn other child
processes.