Add an implementation of call_compiler that uses nvrtc.
This is 2-3x times faster for a simple program that compiles a lot.

Also for the nvrtc compiler to work, I had to simplify a bit how we
define the types since including files is hard. This may lead to a
slight speed-up while using plain NVCC.

While there I added a way for compilers to output theirs logs and
removed the obsolete cuda_set_compiler function.