Numerous improvements to integer factorizer.
- Average-case factorization and margin massively improved. The
factorizer now has optimal 2-, 3- and 5-smooth factorizers and will
always attempt to use them. It will unconditionally try the optimal 5-
smooth factorizer first, even when k>5 and/or slack is unlimited, usually
resulting in a factorization with a product <1% greater than the
original number. This compares with the just-under-100% growth for a
number of the form 2^m+1 with 100% slack.
- Permit a prime x during factorization to attempt decreasing its value
first, then incrementing it.
- New testcases and extra checks in factorization tests.
- New testcases for primality tester.